+++ test/BNO055.cpp
@@ -1,35 +1,133 @@
+#include <Wire.h>
+#include <Adafruit_Sensor.h>
+#include <Adafruit_BNO055.h>
+#include <utility/imumaths.h>
+
+/* This driver reads raw data from the BNO055
+
+   Connections
+   ===========
+   Connect SCL to analog 5
+   Connect SDA to analog 4
+   Connect VDD to 3.3V DC
+   Connect GROUND to common ground
+
+   History
+   =======
+   2015/MAR/03  - First release (KTOWN)
+*/
 
 //マルチスレッドのタスクハンドル格納用です。
 //タスクハンドルはアドレスで渡す必要があるため、ここで配列として設定しています。
 TaskHandle_t thp[1];
 
-//カウントを加算し、シリアルモニタで表示するための変数を設定しています。
-//ここで宣言した変数は、コアを跨いでアクセスできる変数になります。
-long a = 0;
+/* Set the delay between fresh samples */
+#define BNO055_SAMPLERATE_DELAY_MS (100)
 
-//セットアップ関数です。スレッドの設定はここで行います。
-void setup() {
+// Check I2C device address and correct line below (by default address is 0x29 or 0x28)
+//                                   id, address
+Adafruit_BNO055 bno = Adafruit_BNO055(-1, 0x28, &Wire);
+int led = 23;
 
-  //シリアルモニタ表示用の宣言です。
+/**************************************************************************/
+/*
+    Arduino setup function (automatically called at startup)
+*/
+/**************************************************************************/
+void setup(void)
+{
+  pinMode(led, OUTPUT);
   Serial.begin(115200);
+
+  while (!Serial) delay(10);  // wait for serial port to open!
+
+  Serial.println("Orientation Sensor Raw Data Test"); Serial.println("");
+
+  /* Initialise the sensor */
+  if(!bno.begin())
+  {
+    /* There was a problem detecting the BNO055 ... check your connections */
+    Serial.print("Ooops, no BNO055 detected ... Check your wiring or I2C ADDR!");
+    while(1);
+  }
+
+  delay(1000);
+
+  /* Display the current temperature */
+  int8_t temp = bno.getTemp();
+  Serial.print("Current Temperature: ");
+  Serial.print(temp);
+  Serial.println(" C");
+  Serial.println("");
+
+  bno.setExtCrystalUse(true);
+
+  Serial.println("Calibration status values: 0=uncalibrated, 3=fully calibrated");
 
   //ここでスレッドを立てることを宣言しています。
   xTaskCreatePinnedToCore(Core0a, "Core0a", 4096, NULL, 3, &thp[0], 0); 
-  //xTaskCreatePinnedToCore()がスレッドの宣言です。
-  //内容は([タスク名], "[タスク名]", [スタックメモリサイズ(4096or8192)],
-  //      NULL, [タスク優先順位](1-24,大きいほど優先順位が高い)],
-  //      [宣言したタスクハンドルのポインタ(&thp[0])], [Core ID(0 or 1)]); 
 }
 
-void loop() {//メインCPU(Core1)で実行するプログラム
-  Serial.println(a);//変数aの中身をシリアル出力
-  delay(100);//100/1000秒待つ
+/**************************************************************************/
+/*
+    Arduino loop function, called once 'setup' is complete (your own code
+    should go here)
+*/
+/**************************************************************************/
+void loop(void)
+{
+  // Possible vector values can be:
+  // - VECTOR_ACCELEROMETER - m/s^2
+  // - VECTOR_MAGNETOMETER  - uT
+  // - VECTOR_GYROSCOPE     - rad/s
+  // - VECTOR_EULER         - degrees
+  // - VECTOR_LINEARACCEL   - m/s^2
+  // - VECTOR_GRAVITY       - m/s^2
+  imu::Vector<3> euler = bno.getVector(Adafruit_BNO055::VECTOR_EULER);
+
+  /* Display the floating point data */
+  Serial.print("X: ");
+  Serial.print(euler.x());
+  Serial.print(" Y: ");
+  Serial.print(euler.y());
+  Serial.print(" Z: ");
+  Serial.print(euler.z());
+  Serial.print("\t\t");
+
+  /*
+  // Quaternion data
+  imu::Quaternion quat = bno.getQuat();
+  Serial.print("qW: ");
+  Serial.print(quat.w(), 4);
+  Serial.print(" qX: ");
+  Serial.print(quat.x(), 4);
+  Serial.print(" qY: ");
+  Serial.print(quat.y(), 4);
+  Serial.print(" qZ: ");
+  Serial.print(quat.z(), 4);
+  Serial.print("\t\t");
+  */
+
+  /* Display calibration status for each sensor. */
+  uint8_t system, gyro, accel, mag = 0;
+  bno.getCalibration(&system, &gyro, &accel, &mag);
+  Serial.print("CALIBRATION: Sys=");
+  Serial.print(system, DEC);
+  Serial.print(" Gyro=");
+  Serial.print(gyro, DEC);
+  Serial.print(" Accel=");
+  Serial.print(accel, DEC);
+  Serial.print(" Mag=");
+  Serial.println(mag, DEC);
+
+  delay(BNO055_SAMPLERATE_DELAY_MS);
 }
 
 void Core0a(void *args) {//サブCPU(Core0)で実行するプログラム
   while (1) {//ここで無限ループを作っておく
-    //サブで実行するプログラムを書く
-    delay(1);//1/1000秒待つ
-    a++;//aに１を加算する
+    digitalWrite(led, HIGH);
+    delay(500);
+    digitalWrite(led, LOW);
+    delay(500);
   }
 }
